refactor(communities): Making community display page only for selection
