+++ src/modules/communities/pages/CommunitiesPages/components/CommunityCard/CommunityCard.styles.ts
