+++ src/routes/Router.test.tsx
@@ -5,7 +5,11 @@ import Route from './Route';
 describe('route redirections based on authentication', () => {
   it('should render public page', async () => {
     const { getByText } = renderWithRouterAndAuth(
-      <Route path="/dashboard" comp={() => <div>DashboardPage</div>} />,
+      <Route
+        path="/dashboard"
+        isPrivate={false}
+        comp={() => <div>DashboardPage</div>}
+      />,
       { route: '/dashboard' },
     );
 
@@ -15,26 +19,26 @@ describe('route redirections based on authentication', () => {
   it('should redirect to login page when not logged and try to access private page', async () => {
     const { history } = renderWithRouterAndAuth(
       <Route
-        path="/comunidades"
+        path="/selecao-de-comunidades"
         isPrivate
         comp={() => <div>CommunityPage</div>}
       />,
-      { route: '/comunidades' },
+      { route: '/selecao-de-comunidades' },
     );
 
     expect(history.location.pathname).toEqual('/login');
   });
 
-  it('should render private page page when user is logged', async () => {
+  it('should render private page when user is logged', async () => {
     setUserAuthenticated();
 
     const { getByText } = renderWithRouterAndAuth(
       <Route
-        path="/comunidades"
+        path="/selecao-de-comunidades"
         isPrivate
         comp={() => <div>CommunityPage</div>}
       />,
-      { route: '/comunidades' },
+      { route: '/selecao-de-comunidades' },
     );
 
     await waitFor(() => expect(getByText('CommunityPage')).toBeInTheDocument());
